+++ R/MeasureSurvBeggC.R
@@ -22,6 +22,8 @@ MeasureSurvBeggC = R6Class("MeasureSurvBeggC",
   public = list(
     #' @description Creates a new instance of this [R6][R6::R6Class] class.
     initialize = function() {
+      warning("This is now deprecated, use MeasureSurvCindex instead with the
+              function defaults.")
       super$initialize(
         id = "surv.beggC",
         range = 0:1,