+++ R/MeasureSurvIntLogloss.R
@@ -60,10 +60,10 @@ MeasureSurvIntLogloss = R6::R6Class("MeasureSurvIntLogloss",
   private = list(
     .eps = numeric(0),
     .score = function(prediction, ...) {
-      integrated_score(score = weighted_logloss(truth = prediction$truth,
-                                                distribution = prediction$distr,
-                                                times = self$times,
-                                                eps = self$eps),
+      integrated_score(score = intslogloss(truth = prediction$truth,
+                                           distribution = prediction$distr,
+                                           times = self$times,
+                                           eps = self$eps),
                        integrated = self$integrated,
                        method = self$method)
     }