+++ R/MeasureSurvIntLoglossSE.R
@@ -53,10 +53,10 @@ MeasureSurvIntLoglossSE = R6::R6Class("MeasureSurvIntLoglossSE",
     private = list(
       .eps = numeric(0),
       .score = function(prediction, ...) {
-        integrated_se(score = weighted_logloss(truth = prediction$truth,
-                                               distribution = prediction$distr,
-                                               times = self$times,
-                                               eps = self$eps),
+        integrated_se(score = intslogloss(truth = prediction$truth,
+                                          distribution = prediction$distr,
+                                          times = self$times,
+                                          eps = self$eps),
                       integrated = self$integrated)
       }
     )